% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-glm.R
\name{adjust.GLMModel}
\alias{adjust.GLMModel}
\title{Perform GLM adjustment, based on the classes
of the model. Will perform adjustment based on the linear
model type of `model` and also do G-computation or AIPW
based on the second model type of `model`.}
\usage{
\method{adjust}{GLMModel}(model, data)
}
\arguments{
\item{model}{Object of class GLMModel}
}
\description{
Perform GLM adjustment, based on the classes
of the model. Will perform adjustment based on the linear
model type of `model` and also do G-computation or AIPW
based on the second model type of `model`.
}
