% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-linear.R
\name{adjust.LinModel}
\alias{adjust.LinModel}
\title{Perform adjustment for linear models, including ANOVA, ANCOVA,
and ANHECOVA, with or without covariate-adaptive randomization.}
\usage{
\method{adjust}{LinModel}(model, data)
}
\arguments{
\item{model}{A LinModel object and of class ANOVA, ANCOVA, or ANHECOVA}

\item{data}{A RobinDataLinear object}
}
\description{
Perform adjustment for linear models, including ANOVA, ANCOVA,
and ANHECOVA, with or without covariate-adaptive randomization.
}
