% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-SL.R
\name{adjust.SLModel}
\alias{adjust.SLModel}
\title{Perform GLM adjustment, based on the classes
of the model. Will perform adjustment based on the linear
model type of `model` and also do G-computation or AIPW
based on the second model type of `model`.}
\usage{
\method{adjust}{SLModel}(model, data)
}
\description{
Perform GLM adjustment, based on the classes
of the model. Will perform adjustment based on the linear
model type of `model` and also do G-computation or AIPW
based on the second model type of `model`.
}
