% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simdata-car.R
\name{car_sr}
\alias{car_sr}
\title{Generate simple randomization treatment assignments}
\usage{
car_sr(n, p_trt)
}
\arguments{
\item{n}{Number of observations}

\item{p_trt}{Proportion allotted to treatment}
}
\value{
A vector of treatment assignments as 0's and 1's based on simple randomization.
}
\description{
Generate simple randomization treatment assignments
}
\examples{
car_sr(10, p_trt=0.4)

}
