% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-glm.R
\name{descript.GLMModelResult}
\alias{descript.GLMModelResult}
\title{Generate description for glm model result}
\usage{
\method{descript}{GLMModelResult}(x, ...)
}
\arguments{
\item{x}{A GLMModelResult object}

\item{...}{Additional arguments}
}
\description{
Generate description for glm model result
}
