% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-linear.R
\name{descript.LinModelResult}
\alias{descript.LinModelResult}
\title{Generate description for linear model result}
\usage{
\method{descript}{LinModelResult}(x, ...)
}
\arguments{
\item{x}{A LinModelResult object}

\item{...}{Additional arguments}
}
\description{
Generate description for linear model result
}
