% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-adjust.R
\name{.center.dmat}
\alias{.center.dmat}
\title{Get design matrix for the specified adjustment variables
using the data stored.}
\usage{
.center.dmat(dmat)
}
\description{
Get design matrix for the specified adjustment variables
using the data stored.
}
