% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{.df.toclass}
\alias{.df.toclass}
\title{Takes a data frame and converts it to a list with
the attributes as specified by the names passed to ...}
\usage{
.df.toclass(df, classname, ...)
}
\arguments{
\item{df}{data.frame with columns to extract}

\item{classname}{name of class}

\item{...}{Additional names of columns to extract, or vector of names,
e.g., treat_col="treatment", or strata_cols=c("s_1", "s_2")}
}
\description{
Takes a data frame and converts it to a list with
the attributes as specified by the names passed to ...
}
