% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{.edit.formula}
\alias{.edit.formula}
\title{Edits user-specified formula}
\usage{
.edit.formula(form, response_col, treat_col)
}
\arguments{
\item{form}{Formula specified by user, either character or as formula}

\item{response_col}{Name of response column}

\item{treat_col}{Name of treatment column}
}
\description{
# TODO: This needs tests!
}
