% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-adjust.R
\name{.get.dmat}
\alias{.get.dmat}
\title{Get design matrix for the specified adjustment variables
using the data stored.}
\usage{
.get.dmat(data, adj_vars)
}
\description{
Get design matrix for the specified adjustment variables
using the data stored.
}
