% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-adjust.R
\name{.get.vcovHC}
\alias{.get.vcovHC}
\title{Gets the vcovHC weights for the sample size and number of parameters}
\usage{
.get.vcovHC(vcovHC, n, p)
}
\arguments{
\item{vcovHC}{vcovHC option}

\item{n}{Number of data points}

\item{p}{Number of coefficients}
}
\description{
Gets the vcovHC weights for the sample size and number of parameters
}
