% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{.make.model.RoboDataGLM}
\alias{.make.model.RoboDataGLM}
\title{Makes a model class for the specified glm adjustment method
with settings for covariate randomization
scheme and vcovHC type.}
\usage{
\method{.make.model}{RoboDataGLM}(
  data,
  adj_method,
  car_scheme,
  vcovHC,
  covariate_to_include_strata,
  g_family,
  g_accuracy
)
}
\arguments{
\item{adj_method}{Type of adjustment method}

\item{car_scheme}{Type of covariate adaptive randomization}

\item{vcovHC}{Type of heteroskedasticity-consistent SE's}

\item{covariate_to_include_strata}{Include strata as covariates}

\item{g_family}{Family for GLM model}

\item{g_accuracy}{Accuracy tolerance for prediction unbiasedness}
}
\description{
Makes a model class for the specified glm adjustment method
with settings for covariate randomization
scheme and vcovHC type.
}
