% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{.make.model.RoboDataLinear}
\alias{.make.model.RoboDataLinear}
\title{Makes a model class for the specified linear adjustment method
with settings for covariate randomization
scheme and vcovHC type.}
\usage{
\method{.make.model}{RoboDataLinear}(data, adj_method, car_scheme, vcovHC, covariate_to_include_strata)
}
\arguments{
\item{adj_method}{Type of adjustment method}

\item{car_scheme}{Type of covariate adaptive randomization}

\item{vcovHC}{Type of heteroskedasticity-consistent SE's}

\item{covariate_to_include_strata}{Include strata as covariates}
}
\description{
Makes a model class for the specified linear adjustment method
with settings for covariate randomization
scheme and vcovHC type.
}
