% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-linear.R
\name{.vcov_sr.B}
\alias{.vcov_sr.B}
\title{Gets the B matrix for ANCOVA models
in the asymptotic variance formula.}
\usage{
.vcov_sr.B(data, mod)
}
\description{
Gets the B matrix for ANCOVA models
in the asymptotic variance formula.
}
