% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-linear.R
\name{.vcov_sr.diag}
\alias{.vcov_sr.diag}
\title{Gets the diagonal sandwich variance component
for all linear models in the asymptotic variance formula.}
\usage{
.vcov_sr.diag(data, mod, residual = NULL)
}
\description{
Gets the diagonal sandwich variance component
for all linear models in the asymptotic variance formula.
}
