% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-linear.R
\name{.vcov_sr.ScriptB}
\alias{.vcov_sr.ScriptB}
\title{Gets the Script B matrix for AN(HE)COVA models
in the asymptotic variance formula.}
\usage{
.vcov_sr.ScriptB(data, model)
}
\description{
Gets the Script B matrix for AN(HE)COVA models
in the asymptotic variance formula.
}
