% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-results.R
\name{format.results}
\alias{format.results}
\title{Function to format results.}
\usage{
\method{format}{results}(labels, estimates, varcov, label_name = "treat")
}
\description{
Function to format results.
}
