% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-linear.R
\name{linmod.ANOVA}
\alias{linmod.ANOVA}
\title{Fits a linear model to use in ANOVA}
\usage{
\method{linmod}{ANOVA}(model, data, family = gaussian, center = TRUE)
}
\arguments{
\item{model}{Object of class LinModel}

\item{data}{Object of class RoboDataLinear}
}
\description{
Fits a linear model to use in ANOVA
}
