% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-linear.R
\name{linmod.CUSTOM}
\alias{linmod.CUSTOM}
\title{Fits a custom GLM model using the user-specified
formula.}
\usage{
\method{linmod}{CUSTOM}(model, data, family = gaussian, center = TRUE)
}
\arguments{
\item{model}{Object of class LinModel}

\item{data}{Object of class RoboDataLinear}
}
\description{
Fits a custom GLM model using the user-specified
formula.
}
