% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-linear.R
\name{linmod}
\alias{linmod}
\title{Fits a linear model with settings based on model and data.
The linear model is not the "true" model, but it's the linear model
that will aid in adjustment for ANCOVA and ANHECOVA, and used
to estimate the treatment effects for ANOVA (equivalent to a sample)
mean per group.}
\usage{
linmod(model, data, family, center = TRUE)
}
\arguments{
\item{model}{Object of class LinModel}

\item{data}{Object of class RoboDataLinear}
}
\description{
Fits a linear model with settings based on model and data.
The linear model is not the "true" model, but it's the linear model
that will aid in adjustment for ANCOVA and ANHECOVA, and used
to estimate the treatment effects for ANOVA (equivalent to a sample)
mean per group.
}
