% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-car.R
\name{omegaz.closure}
\alias{omegaz.closure}
\title{Compute the Omega(Z) matrix
based on the given covariate adaptive randomization scheme.}
\usage{
omegaz.closure(car_scheme)
}
\description{
This function returns a function to obtain the correct matrix
for a specified vector of randomization probabilities pi_t.
}
