% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.CalibrationResult}
\alias{print.CalibrationResult}
\title{Print calibration result}
\usage{
\method{print}{CalibrationResult}(x, ...)
}
\arguments{
\item{x}{A GLMModel result. If you'd like to calibrate a linear
adjustment, use `robincar_glm` instead of `robincar_linear`.}

\item{...}{Additional arguments}
}
\description{
Print calibration result
}
