% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.ContrastResult}
\alias{print.ContrastResult}
\title{Print contrast result}
\usage{
\method{print}{ContrastResult}(x, ...)
}
\arguments{
\item{x}{A ContrastResult object}

\item{...}{Additional arguments}
}
\description{
Print contrast result
}
