% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.GLMModelResult}
\alias{print.GLMModelResult}
\title{Print glm model result}
\usage{
\method{print}{GLMModelResult}(x, ...)
}
\arguments{
\item{x}{A GLMModelResult object}

\item{...}{Additional arguments}
}
\value{
Prints the treatment mean estimates (and variances) based on a GLM working model,
along with the settings used. See \link[RobinCar:robincar_glm]{RobinCar::robincar_glm()}.
}
\description{
Print glm model result
}
