% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.LinModelResult}
\alias{print.LinModelResult}
\title{Print linear model result}
\usage{
\method{print}{LinModelResult}(x, ...)
}
\arguments{
\item{x}{A LinModelResult object}

\item{...}{Additional arguments}
}
\description{
Print linear model result
}
