% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.MHResult}
\alias{print.MHResult}
\title{Print MH result}
\usage{
\method{print}{MHResult}(x, ...)
}
\arguments{
\item{x}{A ContrastResult object}

\item{...}{Additional arguments}
}
\value{
Prints estimates (and variances) of treatment contrasts based on MH risk difference or ATE
}
\description{
Print MH result
}
