% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.TTEResult}
\alias{print.TTEResult}
\title{Print TTE result}
\usage{
\method{print}{TTEResult}(x, ...)
}
\arguments{
\item{x}{A TTEResult object}

\item{...}{Additional arguments}
}
\description{
Print TTE result
}
