% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunII_ourT_v3.R
\name{RobinCar}
\alias{RobinCar}
\title{Main function for RobinCar}
\usage{
RobinCar(Fun.data, Fun.trt_label.pair, Fun.trt_alc.pair = NA,
  Fun.covariates = NA, Fun.adjmethod = "Hetero")
}
\arguments{
\item{Fun.data}{A dataframe formulated as data.frame(y, I, strata_cross, x1, x2,..., xp). Specifically, y is the observed outcome;
I is the treatment indicator equals 1,...,k; strata_cross is a factor corresponding to the joint levels of Z; x1,...,xp are additional covariates
to be adjusted. Any categorical variable in x1,...,xp needs to be dummied as 0 or 1 (numeric).}

\item{Fun.trt_label.pair}{A vector of length two, e.g. Fun.trt_label.pair=c(s,t), indicates which two treatments are being compared.
The estimand is therefore the average treatment effect comparing the latter to the former.}

\item{Fun.trt_alc.pair}{A vector of length two, e.g., Fun.trt_alc.pair=c(pi_s, pi_t), that corresponds to the target probability of taking treatments s and t, respectively.
If Fun.trt_alc.pair is NA, pi_s and pi_t will be estimated from Fun.data$I.}

\item{Fun.covariates}{A subset of c('x1', ..., 'xp'), which indicates the covariates used for adjustment and should match the column names in Fun.data.
The default is NA, that is no covariate adjustment is made and theta_hat will be applied.}

\item{Fun.adjustment}{Either "Hetero" or "Homo". "Hetero" corresponds to theta_A_hat; "Homo" corresponds to theta_B_hat.
If Fun.covariates is NA, this argument will be omitted. For recommendations of using either "Hetero" or "Homo", please refer to the last section of Ye et al., (2020).}
}
\value{
A list
\describe{
\item{Z_statistic}{Equals \code{estimate}/\code{se}}
\item{estimate}{Estimated average treatment effect}
\item{se}{Standard error of \code{estimate}}
\item{Pval}{Two-sided p-value calculated using the \code{Z_statistic}}
}
}
\description{
Main function for RobinCar
}
\references{
Ting Ye, Yanyao Yi, Jun Shao (2020). Inference on Average Treatment Effect under Minimization and Other Covariate-Adaptive Randomization Methods.
}
