% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-calibrate.R
\name{robincar_calibrate}
\alias{robincar_calibrate}
\title{Run a linear or joint calibration on
a GLMModelResult object.}
\usage{
robincar_calibrate(result, joint = FALSE, add_x = NULL, vcovHC = "HC0")
}
\arguments{
\item{result}{A GLMModelResult}

\item{joint}{If true, then performs joint calibration
with the mu and Z
to achieve universality and efficiency gain
rather than just linear calibration that uses mu.}

\item{add_x}{Additional x to use in the calibration. Must have been in
the original dataset that robincar_glm was called on.}

\item{vcovHC}{Which type of heteroskedasticity-consistent
standard error estimates to use.}
}
\description{
Run a linear or joint calibration on
a GLMModelResult object.
}
