% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-contrast.R
\name{robincar_contrast}
\alias{robincar_contrast}
\title{Create a contrast using a result object (that has both result and varcov)}
\usage{
robincar_contrast(result, contrast_h, contrast_dh = NULL)
}
\arguments{
\item{result}{A LinModelResult or GLMModelResult}

\item{contrast_h}{An optional function to specify a desired contrast}

\item{contrast_dh}{An optional jacobian function for the contrast}
}
\description{
Create a contrast using a result object (that has both result and varcov)
}
