% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-coxscore.R
\name{robincar_coxscore}
\alias{robincar_coxscore}
\title{Robust cox score adjustment}
\usage{
robincar_coxscore(...)
}
\arguments{
\item{...}{Arguments to robincar_tte, other than `adj_method`}
}
\description{
Robust cox score adjustment
}
