% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-glm.R
\name{robincar_glm}
\alias{robincar_glm}
\title{Covariate adjustment using generalized linear working model}
\usage{
robincar_glm(
  df,
  treat_col,
  response_col,
  car_strata_cols = NULL,
  covariate_cols = NULL,
  car_scheme = "simple",
  adj_method = "heterogeneous",
  covariate_to_include_strata = NULL,
  g_family = stats::gaussian,
  g_accuracy = 7,
  formula = NULL,
  contrast_h = NULL,
  contrast_dh = NULL
)
}
\arguments{
\item{df}{A data.frame with the required columns}

\item{treat_col}{Name of column in df with treatment variable}

\item{response_col}{Name of the column in df with response variable}

\item{car_strata_cols}{Names of columns in df with car_strata variables}

\item{covariate_cols}{Names of columns in df with covariate variables}

\item{car_scheme}{Name of the type of covariate-adaptive randomization scheme. One of: "simple", "pocock-simon", "biased-coin", "permuted-block".}

\item{adj_method}{Name of adjustment method to use, one of "heterogeneous" (interaction model) or "homogeneous"}

\item{covariate_to_include_strata}{Whether to include car_strata variables in covariate adjustment. Defaults to F for homogeneous; defaults to T for heterogeneous. User may override by passing in this argument.}

\item{g_family}{Family that would be supplied to glm(...), e.g., binomial. If no link specified, will use default link, like behavior in glm.
If you wish to use a negative binomial working model with an unknown dispersion parameter, then use `g_family="nb"`.}

\item{g_accuracy}{Level of accuracy to check prediction un-biasedness.}

\item{formula}{An optional formula to use for adjustment specified using as.formula("..."). This overrides car_strata_cols and covariate_cols.}

\item{contrast_h}{An optional function to specify a desired contrast}

\item{contrast_dh}{An optional jacobian function for the contrast (otherwise use numerical derivative)}
}
\description{
WARNING: this function is still under development.
Estimate treatment-group-specific response means and (optionally)
treatment group contrasts using a generalized linear working model.
}
