% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-logrank.R
\name{robincar_logrank}
\alias{robincar_logrank}
\title{Robust (potentially stratified) logrank adjustment}
\usage{
robincar_logrank(adj_method, ...)
}
\arguments{
\item{adj_method}{Adjustment method, one of "CL", "CSL"}

\item{df}{A data.frame with the required columns}

\item{treat_col}{Name of column in df with treatment variable}

\item{response_col}{Name of the column in df with response variable}

\item{event_col}{Name of column in df with event indicator
(0/FALSE=no event, 1/TRUE=event)}

\item{strata_cols}{Names of columns in df with strata variables}

\item{covariate_cols}{Names of columns in df with covariate variables}

\item{car_scheme}{Name of the type of covariate-adaptive randomization scheme. One of: "simple", "pocock-simon", "biased-coin", "permuted-block".}

\item{ref_arm}{Reference arm of the treatment group, defaults to NULL,
which results in using the first element of `unique(data[, treat_col])`.}

\item{p_trt}{Treatment allocation ratio for the reference arm.}
}
\description{
Robust (potentially stratified) logrank adjustment
}
\examples{
library(magrittr)
library(dplyr)
set.seed(0)
n=100
data.simu0=data_gen(n=n,
                    theta=0,
                    randomization="permuted_block",
                    p_trt=0.5,
                    case="case2") \%>\% mutate(strata1=sample(letters[1:3],n,replace=TRUE),
                                             strata2=sample(LETTERS[4:5],n,replace=TRUE))

out <- robincar_logrank(df=data.simu0,
                        treat_col="I1",
                        p_trt=0.5,
                        ref_arm=0,
                        response_col="t",
                        event_col="delta",
                        covariate_cols=c("model_z1", "model_z2"),
                        car_scheme="simple",
                        adj_method=c("CL"))

set.seed(0)
n=100
data.simu0=data_gen(n=n,
                    theta=0,
                    randomization="permuted_block",
                    p_trt=0.5,
                    case="case1")

data.simu <- data.simu0 \%>\%
  tidyr::pivot_longer(cols=starts_with("strata"),
                      names_prefix="strata",
                      names_to="strt") \%>\%
  filter(value==1) \%>\% select(-value) \%>\%
  mutate(strt=forcats::as_factor(strt)) \%>\%
  select(t,strt) \%>\%
  left_join(data.simu0, .)

out1 <- robincar_logrank(df=data.simu,
                         treat_col="I1",
                         p_trt=0.5,
                         ref_arm=0,
                         response_col="t",
                         event_col="delta",
                         strata_cols="strt",
                         covariate_cols=NULL,
                         car_scheme=c("permuted-block"),
                         adj_method=c("CSL")
)
}
