% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-SL.R
\name{robincar_SL}
\alias{robincar_SL}
\title{Covariate adjustment using working models from the super learner libraries
through the AIPW package with cross-fitting.
Estimate treatment-group-specific response means and (optionally)
treatment group contrasts using a generalized linear working model.}
\usage{
robincar_SL(
  df,
  treat_col,
  response_col,
  strata_cols = NULL,
  covariate_cols = NULL,
  car_scheme = "simple",
  vcovHC = "HC0",
  covariate_to_include_strata = NULL,
  SL_libraries = c(),
  k_split = 2,
  contrast_h = NULL,
  contrast_dh = NULL
)
}
\arguments{
\item{df}{A data.frame with the required columns}

\item{treat_col}{Name of column in df with treatment variable}

\item{response_col}{Name of the column in df with response variable}

\item{strata_cols}{Names of columns in df with strata variables}

\item{covariate_cols}{Names of columns in df with covariate variables}

\item{car_scheme}{Name of the type of covariate-adaptive randomization scheme. One of: "simple", "pocock-simon", "biased-coin", "permuted-block".}

\item{vcovHC}{Type of heteroskedasticity-consistent variance estimates. One of: "HC0", "HC1", "HC3".}

\item{covariate_to_include_strata}{Whether to include strata variables in covariate adjustment. Defaults to F for ANOVA and ANCOVA; defaults to T for ANHECOVA. User may override by passing in this argument.}

\item{SL_libraries}{Vector of super-learner libraries to use for the covariate adjustment (see SuperLearner::listWrappers())}

\item{k_split}{Number of splits to use in cross-fitting (see ?AIPW) (default is 2)}

\item{contrast_h}{An optional function to specify a desired contrast}

\item{contrast_dh}{An optional jacobian function for the contrast (otherwise use numerical derivative)}

\item{centering}{WARNING: Advanced use only. By default is NULL, so will perform expected behavior
         in documentation.
If centering = 'none', then this will force it to be a g-computation estimator,
   even if prediction unbiasedness does not hold.
If centering = 'tx', then this will center within treatment groups, which is the
   AIPW estimator (equivalent to g-computation if prediction unbiasedness holds).
TODO: If centering = 'tx-strata', then this will center within treatment group and
   within joint strata, which allows for the AIPW estimator to be used with Pocock-Simon.}
}
\description{
Covariate adjustment using working models from the super learner libraries
through the AIPW package with cross-fitting.
Estimate treatment-group-specific response means and (optionally)
treatment group contrasts using a generalized linear working model.
}
