% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-SL.R
\name{robincar_SL_median}
\alias{robincar_SL_median}
\title{Covariate adjustment using working models from the super learner libraries
through the AIPW package with cross-fitting.
Estimate treatment-group-specific response means and (optionally)
treatment group contrasts using a generalized linear working model.}
\usage{
robincar_SL_median(n_times, seed, ...)
}
\arguments{
\item{n_times}{Number of times to run the robincar_SL function}

\item{seed}{Seed to set before running the set of functions}

\item{...}{Additional arguments passed to robincar_SL}
}
\description{
Covariate adjustment using working models from the super learner libraries
through the AIPW package with cross-fitting.
Estimate treatment-group-specific response means and (optionally)
treatment group contrasts using a generalized linear working model.
}
