% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-tte.R
\name{robincar_tte}
\alias{robincar_tte}
\title{Covariate adjustment using generalized linear working model}
\usage{
robincar_tte(
  df,
  treat_col,
  response_col,
  event_col,
  adj_method,
  strata_cols = NULL,
  covariate_cols = NULL,
  p_trt = 0.5,
  ref_arm = NULL,
  sparse_remove = TRUE,
  car_scheme = "simple"
)
}
\arguments{
\item{adj_method}{Adjustment method (one of "CL", "CSL", or "coxscore")}

\item{inheritParams}{robincar_coxscore}
}
\description{
WARNING: this function is still under development.
Estimate treatment-group-specific response means and (optionally)
treatment group contrasts using a generalized linear working model.
}
