% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{validate.RoboDataGLM}
\alias{validate.RoboDataGLM}
\title{Validate GLM model data}
\usage{
\method{validate}{RoboDataGLM}(data)
}
\arguments{
\item{data}{Object of class RoboDataGLM}
}
\description{
Validate GLM model data
}
