% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{validate.RoboDataLinear}
\alias{validate.RoboDataLinear}
\title{Validate linear model data}
\usage{
\method{validate}{RoboDataLinear}(data)
}
\arguments{
\item{data}{Object of class RoboDataLinear}
}
\description{
Validate linear model data
}
