% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{validate.RoboDataTTE}
\alias{validate.RoboDataTTE}
\title{Validate time-to-event data}
\usage{
\method{validate}{RoboDataTTE}(data)
}
\arguments{
\item{data}{Object of class RoboDataTTE}
}
\description{
Validate time-to-event data
}
