% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-linear.R
\name{vcov_car.GLMModel}
\alias{vcov_car.GLMModel}
\title{Gets AIPW asymptotic variance under simple randomization}
\usage{
\method{vcov_car}{GLMModel}(model, data, mod, mutilde)
}
\description{
Gets AIPW asymptotic variance under simple randomization
}
