% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-linear.R
\name{vcov_sr.ANCOVA}
\alias{vcov_sr.ANCOVA}
\title{Gets ANCOVA asymptotic variance under simple randomization}
\usage{
\method{vcov_sr}{ANCOVA}(model, data, mod)
}
\description{
Gets ANCOVA asymptotic variance under simple randomization
}
