% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-linear.R
\name{vcov_sr.ANOVA}
\alias{vcov_sr.ANOVA}
\title{Gets ANOVA asymptotic variance under simple randomization}
\usage{
\method{vcov_sr}{ANOVA}(model, data, mod)
}
\description{
Gets ANOVA asymptotic variance under simple randomization
}
