% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-linear.R
\name{vcov_sr}
\alias{vcov_sr}
\title{Generic function for getting the asymptotic variance-covariance
matrix under simple randomization.}
\usage{
vcov_sr(model, data, mod)
}
\description{
Generic function for getting the asymptotic variance-covariance
matrix under simple randomization.
}
