% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.ContrastResult}
\alias{print.ContrastResult}
\title{Print contrast result}
\usage{
\method{print}{ContrastResult}(x, ...)
}
\arguments{
\item{x}{A ContrastResult object}

\item{...}{Additional arguments}
}
\value{
Prints estimates (and variances) of treatment contrasts based on a linear or GLM working model,
along with the settings used. See \link[RobinCar:robincar_contrast]{RobinCar::robincar_contrast()}
}
\description{
Print contrast result
}
