% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.LinModelResult}
\alias{print.LinModelResult}
\title{Print linear model result}
\usage{
\method{print}{LinModelResult}(x, ...)
}
\arguments{
\item{x}{A LinModelResult object}

\item{...}{Additional arguments}
}
\value{
Prints the treatment mean estimates (and variances) based on a linear working model,
along with the settings used. See \link[RobinCar:robincar_linear]{RobinCar::robincar_linear()}.
}
\description{
Print linear model result
}
