% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-coxscore.R
\name{robincar_coxscore}
\alias{robincar_coxscore}
\title{Robust cox score adjustment}
\usage{
robincar_coxscore(...)
}
\arguments{
\item{df}{A data.frame with the required columns}

\item{treat_col}{Name of column in df with treatment variable}

\item{response_col}{Name of the column in df with response variable}

\item{event_col}{Name of column in df with event indicator
(0/FALSE=no event, 1/TRUE=event)}

\item{strata_cols}{Names of columns in df with strata variables}

\item{covariate_cols}{Names of columns in df with covariate variables}

\item{car_scheme}{Name of the type of covariate-adaptive randomization scheme. One of: "simple", "pocock-simon", "biased-coin", "permuted-block".}

\item{ref_arm}{Reference arm of the treatment group, defaults to NULL,
which results in using the first element of `unique(data[, treat_col])`.}

\item{p_trt}{Treatment allocation ratio for the reference arm.}
}
\description{
Robust cox score adjustment
}
