% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-coxscore.R
\name{robincar_coxscore}
\alias{robincar_coxscore}
\title{Robust cox score adjustment}
\usage{
robincar_coxscore(...)
}
\arguments{
\item{...}{Arguments to robincar_tte, other than `adj_method`}
}
\value{
A result object with the following attributes:

\item{result}{A list: "statistic" is the robust Cox score test statistic which can be used to obtain p-values; "U" and "se" are the numerator and denominator of the test statistic, respectively.}
\item{settings}{The covariate adjustment settings used.}
\item{original_df}{The dataset supplied by the user.}
}
\description{
Robust cox score adjustment
}
