% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robincar-SL.R
\name{robincar_SL_median}
\alias{robincar_SL_median}
\title{BETA: Covariate adjustment using working models from the super learner libraries
through the AIPW package with cross-fitting, with median adjustment.}
\usage{
robincar_SL_median(n_times, seed, ...)
}
\arguments{
\item{n_times}{Number of times to run the robincar_SL function}

\item{seed}{Seed to set before running the set of functions}
}
\description{
Estimate treatment-group-specific response means and (optionally)
treatment group contrasts using a generalized linear working model.
Perform median adjustment to limit randomness induced from cross-fitting.
}
\details{
*WARNING: This function is still under development and has not been extensively tested.*
This function currently only works for two treatment groups.
}
