% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic-tte.R
\name{ttelogic}
\alias{ttelogic}
\title{Logic for performing robincar_coxscore or robincar_logrank for time to event endpoints}
\usage{
ttelogic(x_exists, z_exists, car_scheme, adj_method)
}
\value{
warning or error messages based on the user input,
        whether or not to adjust for covariates, and whether
        to account for covariate-adaptive randomization
}
\description{
Logic for performing robincar_coxscore or robincar_logrank for time to event endpoints
}
\examples{
logic_CL <- ttelogic(x_exists=TRUE, z_exists=TRUE,
                              car_scheme="permuted-block",
                              adj_method="CL")

logic_CSL <- ttelogic(x_exists=TRUE, z_exists=TRUE,
                               car_scheme="permuted-block",
                               adj_method="CSL")

logic_coxscore <- ttelogic(x_exists=TRUE, z_exists=TRUE,
                                    car_scheme="permuted-block",
                                    adj_method="coxscore")
}
